restuctured code
